refactor: createVoucher 반환타입 지정
